requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.4
boto3==1.34.115
django-storages==1.14.2
Pillow==10.3.0
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0
whitenoise==6.7.0
openai==1.45.0
httpx==0.27.2
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
uvicorn[standard]==0.30.0
